% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloclize-package.R
\name{make_flames_venv}
\alias{make_flames_venv}
\title{make FLAMES python package venv}
\usage{
make_flames_venv(
  environment_name = "coloclize_flames_env",
  environment_path = tempdir()
)
}
\arguments{
\item{environment_name}{name of the environment to create}

\item{environment_path}{path to create the environment in}
}
\value{
list with environment name and path
}
\description{
make FLAMES python package venv
}
